+++ sdks/python/apache_beam/io/gcp/bigquery_read_it_test.py
@@ -160,15 +160,9 @@ class ReadTests(BigQueryReadIntegrationTests):
     query = StaticValueProvider(str, self.query)
     with beam.Pipeline(argv=self.args) as p:
       result = (
-          p | 'read' >> beam.io._ReadFromBigQuery(
-              query=self.query, use_standard_sql=True, project=self.project))
-      assert_that(result, equal_to(self.TABLE_DATA))
-
-    with beam.Pipeline(argv=self.args) as p:
-      vp_result = (
-          p | 'read value provider query' >> beam.io._ReadFromBigQuery(
+          p | 'read with value provider query' >> beam.io._ReadFromBigQuery(
               query=query, use_standard_sql=True, project=self.project))
-      assert_that(vp_result, equal_to(self.TABLE_DATA))
+      assert_that(result, equal_to(self.TABLE_DATA))
 
 
 class ReadNewTypesTests(BigQueryReadIntegrationTests):